Root execute by script Added screencap support for snapshot

+++ RCSAndroid/src/com/android/deviceinfo/file/AutoFile.java
@@ -345,19 +345,16 @@ public final class AutoFile {
 		return file.getName();
 	}
 
-	/**
-	 * Flush.
-	 */
-	@Deprecated
-	public void flush() {
-
-	}
 
 	/**
 	 * Delete the file.
 	 */
-	public void delete() {
-		file.delete();
+	public boolean delete() {
+		if(file.exists()){
+			file.delete();
+			return true;
+		}
+		return false;
 	}
 
 	public boolean dropExtension(String ext) {
